refactor: slim section-tree and section-address to format-agnostic core
Why: buildSectionTree operates on mdast — inherently markdown-specific — yet
lived in section-tree.ts alongside genuinely format-agnostic helpers. Same for
parseAddress in section-address.ts (strips markdown heading markers). Leaving
them in "shared" modules blocks a clean core/en-quire boundary.

What changed:
- buildSectionTree (+ extractHeadings, findBodyEnd) moved from section-tree.ts
  into markdown-parser.ts.
- parseAddress (+ stripAddressMarkers) moved from section-address.ts into
  markdown-parser.ts.
- section-tree.ts now holds only format-agnostic helpers: fixSectionEndOffsets
  (newly exported), buildPreambleNode, getBreadcrumb, getSectionPath,
  flattenTree.
- section-address.ts keeps only generic resolvers (resolveAddress,
  resolveSingleSection, etc.).
- Dropped dead import of buildSectionTree from section-ops.ts.
- Dropped unused parseMarkdownAddress import from yaml-parser.ts.
- Test imports updated; no behaviour change (361/361 green).

Step 3 of the en-core/en-quire/en-scribe monorepo split (issue #58).
