+++ src/document/section-ops.ts
@@ -7,7 +7,7 @@ import type {
   FindReplaceOptions,
 } from '../shared/types.js';
 import { parseMarkdown } from './parser.js';
-import { buildSectionTree, getSectionPath, getBreadcrumb, flattenTree } from './section-tree.js';
+import { getSectionPath, getBreadcrumb, flattenTree } from './section-tree.js';
 import { resolveSingleSection, resolveAddress } from './section-address.js';
 import { countCodePoints, offsetToLine } from './ast-utils.js';
 import { countWords } from '../shared/word-count.js';
