+++ test/unit/document/i18n.test.ts
@@ -4,7 +4,8 @@ import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import Database from 'better-sqlite3';
 import { parseMarkdown } from '../../../src/document/parser.js';
-import { buildSectionTree, getSectionPath, flattenTree } from '../../../src/document/section-tree.js';
+import { buildSectionTree } from '../../../src/document/markdown-parser.js';
+import { getSectionPath, flattenTree } from '../../../src/document/section-tree.js';
 import { resolveAddress } from '../../../src/document/section-address.js';
 import { readSection, buildOutline, findReplace } from '../../../src/document/section-ops.js';
 import { initSearchSchema } from '../../../src/search/schema.js';
