+++ test/unit/document/section-tree.test.ts
@@ -3,7 +3,8 @@ import { describe, it, expect } from 'vitest';
 import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { parseMarkdown } from '../../../src/document/parser.js';
-import { buildSectionTree, getBreadcrumb, getSectionPath, flattenTree } from '../../../src/document/section-tree.js';
+import { buildSectionTree } from '../../../src/document/markdown-parser.js';
+import { getBreadcrumb, getSectionPath, flattenTree } from '../../../src/document/section-tree.js';
 
 const fixturesDir = resolve(import.meta.dirname, '../../fixtures/docs');
