+++ test/unit/security/regex-flags.test.ts
@@ -3,7 +3,7 @@ import { describe, it, expect } from 'vitest';
 import { readFileSync } from 'node:fs';
 import { resolve } from 'node:path';
 import { parseMarkdown } from '../../../src/document/parser.js';
-import { buildSectionTree } from '../../../src/document/section-tree.js';
+import { buildSectionTree } from '../../../src/document/markdown-parser.js';
 import { findReplace } from '../../../src/document/section-ops.js';
 
 const fixturesDir = resolve(import.meta.dirname, '../../fixtures/docs');
